% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAvgRelMAE.R
\name{calculateAvgRelMAE}
\alias{calculateAvgRelMAE}
\title{AvgRelMAE accuracy tables and plots}
\usage{
calculateAvgRelMAE(af, bench, sort = FALSE, digits = 3)
}
\arguments{
\item{af}{dataframe with forecast data formatted using the Actual and Forecast
Table Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{bench}{method_id for the benchmark method}

\item{sort}{if TRUE, sorts accuracy results according to the average AvgRelMAE
rank across all available horizons.}

\item{digits}{number of digits after the dot used to output AvgRelMAE; if digits=NA,
AvgRelMAEs and ranks are not rounded.}
}
\value{
a list with the following variables:
\describe{
  \item{accuracy:}{dataframe with accuracy table for AvgRelMAE}
  \item{mae:}{dataframe with MAEs, RelMAE, logRelMAE for each time series for each horizon}
  \item{rank:}{ dataframe with AvgRelMAE ranks}
  \item{plot:}{ggplot2 object showing accuracy vs horizon for AvgRelMAE}
  \item{boxplot:}{ggplot2 object with boxplots showing logRelMAEs for each method}
}
}
\description{
Generates AvgRelMAE accuracy tables and plots.
}
\details{
Calculates AvgRelMAEs (Davydenko and Fildes, 2013) for each horizon and
each method contained in af. Outputs accuracy tables and plots.

The input dataframe af can be created using \code{createAFTS()}.
}
\examples{
af <- createAFTS(example1_TSTS, example1_FTS)
acc <- calculateAvgRelMAE(af, "A")
acc$accuracy
acc$rank
acc$plot

}
\references{
Davydenko, A., & Fildes, R. (2013). \href{https://forvis.github.io/our-publications/5.pdf}{Measuring Forecasting
Accuracy:The Case of Judgmental Adjustments to Sku-Level Demand Forecasts.}\emph{International Journal of Forecasting,} 29(3), 510-512.

Davydenko, A., & Fildes, R. (2016). \href{https://forvis.github.io/our-publications/2.pdf}{Forecast Error Measures:
Critical Review and Practical Recommendations.}
In \emph{Business Forecasting: Practical Problems and Solutions}. John Wiley & Sons Inc.
}
\seealso{
\code{\link{calculateMAPE}}, \code{\link{calculateSMAPE}}, \code{\link{calculateMASE}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
