% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCoverage.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{Coverage probability evaluation}
\usage{
calculateCoverage(af, pi, conf.level = 0.95)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and
Forecast Table Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{pi}{nominal coverage probability (confidence level for the prediction interval to be assessed).}

\item{conf.level}{confidence level for coverage probability estimates, default 0.95.}
}
\value{
dataframe with coverage probability estimates for each method and each horizon
contained in \code{af}.
}
\description{
Calculates estimates of actual coverage probabilities for the given
nominal coverage probability.
}
\details{
Calculates coverage probability estimates for each method and each horizon contained in \code{af}.
\emph{Coverage probability estimate} is the percentage of cases when prediction interval contained actual.
Confidence limits for the actual \emph{coverage probability} are obtained using binom.test function (two.sided=TRUE).

To get a coverage probability graph, use \code{plotCoverage()}.
}
\examples{
af <- createAFTS(example1_ts, example1_fc)
calculateCoverage(af, 90)

}
\seealso{
\code{\link{plotCoverage}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
