% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMAE.R
\name{calculateMAE}
\alias{calculateMAE}
\title{MAE accuracy tables and plots}
\usage{
calculateMAE(af, sort = FALSE, digits = 3)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and Forecast Table
Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{sort}{if TRUE, sorts accuracy results according to the average MAE
rank across all available horizons.}

\item{digits}{number of digits after the dot used to output MdAPE; if digits=NA,
MAEs and ranks are not rounded.}
}
\value{
a list with the following variables:
\describe{
  \item{accuracy:}{dataframe with accuracy table for MAE}
  \item{rank:}{ dataframe with MAE ranks}
  \item{plot:}{ggplot2 object showing accuracy vs horizon for MAE}
}
}
\description{
Generates MAE accuracy tables and plots.
}
\details{
Calculates MAEs for each horizon and each method contained in af.
Outputs accuracy tables and plots.

The input dataframe af can be created using \code{createAFTS()}.
}
\examples{
af <- createAFTS(example1_ts, example1_fc)
acc <- calculateMAE(af)
acc$accuracy
acc$rank
acc$plot

}
\seealso{
\code{\link{calculateMAPE}}, \code{\link{calculateAvgRelMAE}},
\code{\link{calculateSMAPE}}, \code{\link{calculateMASE}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
