% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMdAPE.R
\name{calculateMdAPE}
\alias{calculateMdAPE}
\title{MdAPE accuracy tables and plots}
\usage{
calculateMdAPE(af, sort = FALSE, digits = 3)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and Forecast Table
Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{sort}{if TRUE, sorts accuracy results according to the average MdAPE
rank across all available horizons.}

\item{digits}{number of digits after the dot used to output MdAPE; if digits=NA,
MdAPEs and ranks are not rounded.}
}
\value{
a list with the following variables:
\describe{
  \item{accuracy:}{dataframe with accuracy table for MdAPE}
  \item{rank:}{ dataframe with MdAPE ranks}
  \item{plot:}{ggplot2 object showing accuracy vs horizon for MdAPE}
}
}
\description{
Generates MdAPE accuracy tables and plots.
}
\details{
Calculates MdAPEs for each horizon and each method contained in af.
Outputs accuracy tables and plots.

The input dataframe af can be created using \code{createAFTS()}.
}
\examples{
af <- createAFTS(example1_ts, example1_fc)
acc <- calculateMdAPE(af)
acc$accuracy
acc$rank
acc$plot

}
\seealso{
\code{\link{calculateMAPE}}, \code{\link{calculateAvgRelMAE}},
\code{\link{calculateSMAPE}}, \code{\link{calculateMASE}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
