% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRMSE.R
\name{calculateRMSE}
\alias{calculateRMSE}
\title{RMSE accuracy tables and plots}
\usage{
calculateRMSE(af, sort = FALSE, digits = 3)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and Forecast Table
Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{sort}{if TRUE, sorts accuracy results according to the average RMSE
rank across all available horizons.}

\item{digits}{number of digits after the dot used to output MdAPE; if digits=NA,
RMSEs and ranks are not rounded.}
}
\value{
a list with the following variables:
\describe{
  \item{accuracy:}{dataframe with accuracy table for RMSE}
  \item{rank:}{ dataframe with RMSE ranks}
  \item{plot:}{ggplot2 object showing accuracy vs horizon for RMSE}
}
}
\description{
Generates RMSE accuracy tables and plots.
}
\details{
Calculates RMSEs for each horizon and each method contained in af.
Outputs accuracy tables and plots.

The input dataframe af can be created using \code{createAFTS()}.
}
\examples{
af <- createAFTS(example1_ts, example1_fc)
acc <- calculateRMSE(af)
acc$accuracy
acc$rank
acc$plot

}
\seealso{
\code{\link{calculateMAPE}}, \code{\link{calculateAvgRelRMSE}},
\code{\link{calculateSMAPE}}, \code{\link{calculateMASE}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
