% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSMAPE.R
\name{calculateSMAPE}
\alias{calculateSMAPE}
\title{SMAPE accuracy tables and plots}
\usage{
calculateSMAPE(af, sort = FALSE, digits = 3)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and Forecast Table
Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{sort}{if TRUE, sorts accuracy results according to the average SMAPE
rank across all available horizons.}

\item{digits}{number of digits after the dot used to output SMAPE; if digits=NA,
SMAPEs and ranks are not rounded.}
}
\value{
a list with the following variables:
\describe{
  \item{accuracy:}{dataframe with accuracy table for SMAPE}
  \item{rank:}{ dataframe with SMAPE ranks}
  \item{plot:}{ggplot2 object showing accuracy vs horizon for SMAPE}
}
}
\description{
Generates SMAPE accuracy tables and plots.
}
\details{
Calculates SMAPEs for each horizon and each method contained in \code{af}.
Outputs accuracy tables and plots.

The input dataframe \code{af} can be created using \code{createAFTS()}.
}
\examples{
af <- createAFTS(example1_TSTS, example1_FTS)
acc <- calculateSMAPE(af)
acc$accuracy
acc$rank
acc$plot

}
\seealso{
\code{\link{calculateMAPE}}, \code{\link{calculateAvgRelMAE}}, \code{\link{calculateMASE}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
