% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAFTS.R
\name{createAFTS}
\alias{createAFTS}
\title{Create a dataframe formatted using the Actual and Forecast Table Schema (AFTS)}
\usage{
createAFTS(ts, fc)
}
\arguments{
\item{ts}{dataframe containing time series actuals formatted using the Time Series
Table Schema (TSTS), use \code{showTSTS()} to display schema specification details.}

\item{fc}{dataframe containing forecasts formatted using the Forecast Table Schema (FTS),
use \code{showFTS()} to display schema specification details.}
}
\value{
dataframe in the AFTS format, use \code{showAFTS()} to display schema specification details.
}
\description{
Joins a table containing time series actuals (given in the TSTS format) and a
table containing forecasts (given in the FTS format) to create a table containing both
actuals and forecasts using the Actual and Forecast Table Schema (AFTS) format.
}
\details{
Takes all records contained in \code{fc} and matches with values from \code{ts}.
If no matching value is found, the record is not included in the result.
}
\examples{
af <- createAFTS(example1_TSTS, example1_FTS)
head(af)

}
\seealso{
\code{\link{showTSTS}}, \code{\link{showFTS}}, \code{\link{example1_TSTS}}, \code{\link{example1_FTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
