% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_fc}
\alias{example1_fc}
\title{Example Dataset Formatted Using Forecast Table Schema (FTS)}
\format{\code{example1_fc} is a data frame formatted using the Forecast
Table Schema, use \code{showFTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
example1_fc
}
\description{
Example dataset containing forecast data to different series. The dataset contains forecast data
 created to illustrate the capabilities of the package.
}
\examples{
head(example1_fc, 10)



}
\seealso{
\code{\link{m3_yearly_fc}}, \code{\link{m3_yearly_fc_pis}}, \code{\link{m3_quarterly_fc}},
\code{\link{m3_quarterly_fc_pis}}, \code{\link{m3_monthly_fc}}, \code{\link{m3_monthly_fc_pis}},
\code{\link{m3_other_fc}}, \code{\link{m3_other_fc_pis}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
