% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_FTS}
\alias{example1_FTS}
\title{Example Dataset Formatted Using Forecast Table Schema (FTS)}
\format{\code{example1_FTS} is a data frame formatted using the Forecast
Table Schema, use \code{showFTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
example1_FTS
}
\description{
Example dataset containing forecast data to different series. The dataset contains forecast data
 created to illustrate the capabilities of the package.
}
\examples{
head(example1_FTS, 10)



}
\seealso{
\code{\link{example1_TSTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
