% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_ts}
\alias{example1_ts}
\title{Example Dataset Formatted Using Time Series Table Schema (TSTS)}
\format{\code{example1_ts} is a data frame formatted using the Time Series
Table Schema, use \code{showTSTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
example1_ts
}
\description{
Example dataset containing time series actuals relating to different series. The dataset contains artificial
yearly observations created to illustrate the capabilities of the package.
}
\examples{

head(example1_ts, 10)


}
\seealso{
\code{\link{m3_yearly_ts}}, \code{\link{m3_quarterly_ts}}, \code{\link{m3_monthly_ts}},
\code{\link{m3_other_ts}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
