% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_TSTS}
\alias{example1_TSTS}
\title{Example Dataset Formatted Using Time Series Table Schema (TSTS)}
\format{\code{example1_TSTS} is a data frame formatted using the Time Series
Table Schema, use \code{showTSTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
example1_TSTS
}
\description{
Example dataset containing time series actuals relating to different series. The dataset contains artificial
yearly observations created to illustrate the capabilities of the package.
}
\examples{

head(example1_TSTS, 10)


}
\seealso{
\code{\link{example1_FTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
