% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3_monthly_fc}
\alias{m3_monthly_fc}
\title{Monthly forecasts from the M3-competition (FTS)}
\format{\code{m3_monthly_fc} is a data frame formatted using the Forecast
Table Schema, use \code{showFTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
m3_monthly_fc
}
\description{
The dataset contains forecast data to different monthly time series from the M3-competition.
}
\examples{

head(m3_monthly_fc, 10)


}
\seealso{
\code{\link{example1_fc}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
