% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3_monthly_ts}
\alias{m3_monthly_ts}
\title{Monthly actuals from the M3-competition (TSTS)}
\format{\code{m3_monthly_ts} is a data frame formatted using the Time Series
Table Schema, use \code{showTSTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
m3_monthly_ts
}
\description{
The dataset contains monthly time series actuals relating to different series from the M3-competition.
}
\examples{

head(m3_monthly_ts, 10)


}
\seealso{
\code{\link{example1_ts}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
