% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3_yearly_fc_pis}
\alias{m3_yearly_fc_pis}
\title{Yearly forecasts containing prediction intervals calculated for the M3-competition data (FTS)}
\format{\code{m3_yearly_fc_pis} is a data frame formatted using the Forecast
Table Schema, use \code{showFTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
m3_yearly_fc_pis
}
\description{
The dataset contains forecast data containing prediction intervals calculated for different
yearly time series from the M3-competition.
}
\examples{

head(m3_yearly_fc_pis, 10)


}
\seealso{
\code{\link{example1_fc}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
