% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3_yearly_ts}
\alias{m3_yearly_ts}
\title{Yearly actuals from the M3-competition (TSTS)}
\format{\code{m3_yearly_ts} is a data frame formatted using the Time Series
Table Schema, use \code{showTSTS()} to display schema specification details.}
\source{
\url{https://github.com/forvis/forvision_data/}.
}
\usage{
m3_yearly_ts
}
\description{
Example dataset containing time series actuals relating to different series. The dataset contains artificial
yearly observations created to illustrate the capabilities of the package.
}
\examples{

head(m3_yearly_ts, 10)


}
\seealso{
\code{\link{example1_FTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{datasets}
