% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Coverage Probability Plot}
\usage{
plotCoverage(af, pi, methods = NULL, horizons = NULL,
  conf.level = 0.95, coord.flip = FALSE)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and
Forecast Table Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{pi}{nominal coverage probability (confidence level for the prediction interval to be assessed).}

\item{methods}{array containing method_id values for methods to be shown on the graph,
if NULL, all available methods will be shown.}

\item{horizons}{array containing horizons to be shown on the graph,
if NULL, all available horizons will be shown.}

\item{conf.level}{confidence level for coverage probability estimates, default 0.95.}

\item{coord.flip}{use FALSE for horizontal bar chart, TRUE for vertical.}
}
\value{
Returns a ggplot2 object containing the coverage probability plot.
}
\description{
Generates a coverage probability plot showing coverage probability estimates,
corresponding error bands, and nominal coverage probability.
}
\details{
Generates a bar chart showing coverage percentages for the given nominal coverage
(Sai, Davydenko, and Shcherbakov, 2018). The results are confined to methods contained in methods
and horizons contained in horizons. Adds error bars and nominal coverage line to chart.
}
\examples{
af <- createAFTS(example1_ts, example1_fc)
plotCoverage(af, pi = 90, methods = c("A", "B"))
plotCoverage(af, pi = 90, horizons = 1:2)

}
\references{
Sai, C., Davydenko, A., & Shcherbakov, M. (2018, November). \href{https://forvis.github.io/our-publications/1.pdf}{Data schemas for forecasting (with examples in R).}
\emph{Seventh International Conference on System Modelling & Advancement on Research Trends}, 145-149. Moradabad, India.
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{evaluate}
