% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFan.R
\name{plotFan}
\alias{plotFan}
\title{Fan chart}
\usage{
plotFan(ts, fc, series, origin, method, graphLib = "dygraphs",
  piColor = "blue")
}
\arguments{
\item{ts}{dataframe containing time series actuals formatted using the Time Series Table Schema (TSTS),
use \code{showTSTS()} to display schema specification details. \code{ts} must contain \code{timestamp_dbo} column with
timestamps formatted as an appropriate time-based object.}

\item{fc}{dataframe containing forecasts formatted using the Forecast Table Schema (FTS), use \code{showFTS()}
to display schema specification details. \code{fc} must contain \code{timestamp_dbo} column with timestamps formatted
as an appropriate time-based object.}

\item{series}{series_id for the time series to be shown.}

\item{origin}{origin_timestamp for the origin of the forecast to be shown.}

\item{method}{method_id for the method to be shown.}

\item{graphLib}{choose method for disp}
}
\value{
A dygraph object with  the fan chart.
}
\description{
Draws a fan chart for forecasts for a specified method, time series, and origin.
}
\details{
Generates a fan chart joining a simple line chart for actuals (given in the TSTS format) with
ranges for possible future values (given in the FTS format). Uses all the prediction intervals available in fc.

Current time-based objects supported are \code{Date}, \code{POSIXct}, \code{chron},
\code{yearmon}, \code{yearqtr}, and \code{timeDate}.
}
\examples{
# obtain data
ts <- example1_ts
fc <- example1_fc

# create the "timestamp_dbo" column
library(zoo)
ts$timestamp_dbo <- as.yearmon(ts$timestamp, format = '\%Y')
fc$timestamp_dbo <- as.yearmon(fc$timestamp, '\%Y')

# plot the fan chart for some given origin, series, and method
plotFan(ts, fc, "Y1", "1988", "A")
\dontrun{
# produces error because the input dataframes do not contain the "timestamp_dbo" column
plotFan(example1_TSTS, example1_FTS, "Y1", "1988", "A")
}


}
\seealso{
\code{\link{plotFixedHorizon}}, \code{\link{plotFixedOrigin}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{explore}
