% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFixedHorizon.R
\name{plotFixedHorizon}
\alias{plotFixedHorizon}
\title{Fixed horizon graph}
\usage{
plotFixedHorizon(ts, fc = NULL, series, h, method = NULL)
}
\arguments{
\item{ts}{dataframe containing time series actuals formatted using the Time Series Table Schema (TSTS),
use \code{showTSTS()} to display schema specification details. \code{ts} must contain \code{timestamp_dbo} column with
timestamps formatted as an appropriate time-based object.}

\item{fc}{dataframe containing forecasts formatted using the Forecast Table Schema (FTS), use \code{showFTS()}
to display schema specification details. \code{fc} must contain \code{timestamp_dbo} column with timestamps formatted
as an appropriate time-based object. if NULL, forecasts are not shown on the graph.}

\item{series}{series_id for the time series to be shown.}

\item{h}{forecast horizon for the horizon of the forecast to be shown.}

\item{method}{method_id for the method to be shown.}
}
\value{
a dygraph object containing the fixed horizon graph.
}
\description{
Plots a graph showing point forecasts produced for the same time series, with the same horizon,
but from various rolling origins. Forecasts can relate to one method or to a set of alternative methods.
}
\details{
Uses the following algorithm to plot the graph:

1) select all the forecasts from fc having the specified series_id, horizon, and method

2) plot point forecasts for the rows selected, use different colors for different methods

3) plot actuals contained in ts for the given series_id

Current time-based objects supported are \code{Date}, \code{POSIXct}, \code{chron},
\code{yearmon}, \code{yearqtr}, and \code{timeDate}.
}
\examples{
# obtain data
ts <- example1_ts
fc <- example1_fc

# create the "timestamp_dbo" column
library(zoo)
ts$timestamp_dbo <- as.yearmon(ts$timestamp, format = '\%Y')
fc$timestamp_dbo <- as.yearmon(fc$timestamp, '\%Y')

# plot a graph for some given horizon, series, and method
plotFixedHorizon(ts, fc, "Y1", 1, "A")
\dontrun{
# produces error because the input dataframes do not contain the "timestamp_dbo" column
plotFixedHorizon(example1_TSTS, example1_FTS, "Y1", 1, "A")
}

}
\seealso{
\code{\link{plotFan}}, \code{\link{plotFixedOrigin}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim shcherbakov.
}
\keyword{explore}
