% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFixedOrigin.R
\name{plotFixedOrigin}
\alias{plotFixedOrigin}
\title{Fixed origin graph}
\usage{
plotFixedOrigin(ts, fc = NULL, series, origin, method = NULL)
}
\arguments{
\item{ts}{dataframe containing time series actuals formatted using the Time Series Table Schema (TSTS),
use \code{showTSTS()} to display schema specification details. \code{ts} must contain \code{timestamp_dbo} column with
timestamps formatted as an appropriate time-based object.}

\item{fc}{dataframe containing forecasts formatted using the Forecast Table Schema (FTS), use \code{showFTS()}
to display schema specification details. \code{fc} must contain \code{timestamp_dbo} column with timestamps formatted
as an appropriate time-based object. if NULL, forecasts are not shown on the graph.}

\item{series}{series_id for the time series to be shown.}

\item{origin}{origin_timestamp for the origin of the forecast to be shown.}

\item{method}{method_id for the method to be shown.}
}
\value{
a dygraph object containing the fixed origin graph.
}
\description{
Plots the fixed origin graph showing point forecasts produced for the same time series from
the same origin, with different horizons. Forecasts can relate to one method or to a set of alternative methods.
}
\details{
Uses the following algorithm to plot the graph:

1) select all the forecasts from fc having the specified origin and series_id and containing forecasts
produced by methods included in methods,

2) plot point forecasts, use different colors for different methods

3) plot actuals contained in ts for the given series_id.

Current time-based objects supported are \code{Date}, \code{POSIXct}, \code{chron},
\code{yearmon}, \code{yearqtr}, and \code{timeDate}.
}
\examples{
# obtain data
ts <- example1_ts
fc <- example1_fc

# create the "timestamp_dbo" column
library(zoo)
ts$timestamp_dbo <- as.yearmon(ts$timestamp, format = '\%Y')
fc$timestamp_dbo <- as.yearmon(fc$timestamp, '\%Y')

# plot a graph for some given origin, series, and method
plotFixedOrigin(ts, fc, "Y1", "1988", "A")
\dontrun{
# produces error because the input dataframes do not contain the "timestamp_dbo" column
plotFixedOrigin(example1_TSTS, example1_FTS, "Y1", "1988", "A")
}

}
\seealso{
\code{\link{plotFan}}, \code{\link{plotFixedHorizon}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{explore}
