% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRD.R
\name{plotPRD}
\alias{plotPRD}
\title{Prediction-Realization Diagram}
\usage{
plotPRD(af, useLogs = FALSE, xlim = NULL, ylim = NULL)
}
\arguments{
\item{af}{dataframe containing forecast data formatted using the Actual and Forecast Table
Schema (AFTS), use \code{showAFTS()} to display schema specification details.}

\item{useLogs}{if TRUE, uses logarithmic scales.}

\item{xlim}{the minimum and maximum values for x-axis, if NULL, finds min/max values based on the data.}

\item{ylim}{the minimum and maximum values for y-axis, if NULL, finds min/max values based on the data.}
}
\value{
Returns a ggplot2 object.
}
\description{
Generates a prediction-realization diagram showing a scatterplot with forecast vs actual values.
}
\details{
Generates a \href{https://forvis.github.io/our-publications/1.pdf}{prediction-realization diagram} for the given dataset.
The diagram shows a scatter plot with actuals and forecasts.
}
\examples{
af <- createAFTS(example1_TSTS, example1_FTS)
plotPRD(af)

}
\references{
Sai, C., Davydenko, A., & Shcherbakov, M. (2018, November). \href{https://forvis.github.io/our-publications/1.pdf}{Data schemas for forecasting (with examples in R).}
\emph{Seventh International Conference on System Modelling & Advancement on Research Trends}, 145-149. Moradabad, India.
}
\seealso{
\code{\link{plotFixedOrigin}}, \code{\link{plotFan}}, \code{\link{plotFixedHorizon}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
\keyword{explore}
