% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSeriesSummary.R
\name{printSeriesSummary}
\alias{printSeriesSummary}
\title{Time Series Summary Table}
\usage{
printSeriesSummary(ts)
}
\arguments{
\item{ts}{dataframe containing time series actuals formatted using the Times Series Table Schema (TSTS),
use \code{showTSTS()} to display schema specification details.}
}
\description{
Prints summary info for actuals.
}
\details{
Outputs summary info containing the number of series, number of observations, period range.
}
\examples{
printSeriesSummary(example1_TSTS)

}
\seealso{
\code{\link{printForecastSummary}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
