% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGlossary.R
\name{showGlossary}
\alias{showGlossary}
\title{Glossary}
\usage{
showGlossary()
}
\description{
Display definitions for terms used in package documentation.
}
\details{
Opens terms definitions.
}
\seealso{
\code{\link{showTSTS}}, \code{\link{showFTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
