% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateFTS.R
\name{validateFTS}
\alias{validateFTS}
\title{Forecast Table Schema (FTS) Validation}
\usage{
validateFTS(fc)
}
\arguments{
\item{fc}{dataframe containing forecasts formatted using the Forecast Table Schema (FTS),
use \code{showFTS()} to display schema specification details.}
}
\value{
TRUE if the checks are passed, FALSE otherwise.
}
\description{
Checks if the input data is correctly formatted in accordance with the Forecast Table Schema (FTS).
}
\details{
Checks that fc contains necessary column and the composite primary key values are not duplicated.
}
\examples{
validateFTS(example1_FTS)

}
\seealso{
\code{\link{validateTSTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
