% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateTSTS.R
\name{validateTSTS}
\alias{validateTSTS}
\title{Time Series Table Schema (TSTS) Validation}
\usage{
validateTSTS(ts)
}
\arguments{
\item{ts}{dataframe containing actuals formatted using the Time Series Table Schema (TSTS),
use \code{showTSTS()} to display schema specification details.}
}
\value{
TRUE if the checks are passed, FALSE otherwise.
}
\description{
Checks if the input data is correctly formatted in accordance with the Time Series Table Schema (TSTS).
}
\details{
Checks that ts contains necessary column and the composite primary key values are not duplicated.
}
\examples{
validateTSTS(example1_TSTS)

}
\seealso{
\code{\link{validateFTS}}
}
\author{
Cuong Sai, Andrey Davydenko, and Maxim Shcherbakov.
}
